+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### ✅ Added
 - Support for custom message receipt states
 - Scrolling of instant commands
+- Config to turn off tab bar visibility handling
 
 ### 🔄 Changed
 - Updated Nuke dependency to 11.3.0 for SPM